# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go + Gin backend for a multivendor eCommerce platform. Auth, vendor onboarding (KYC), product catalog, order splitting, commission, settlement.",
        "title": "Coolmate Multivendor eCommerce API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Coolmate Team",
            "email": "dev@coolmate.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/vendors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors (Admin)"
                ],
                "summary": "List all vendors (STUB, admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/vendors/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors (Admin)"
                ],
                "summary": "Get vendor by ID (STUB, admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/vendors/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors (Admin)"
                ],
                "summary": "Approve vendor (STUB, admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/vendors/{id}/bank-details": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors (Admin)"
                ],
                "summary": "Update vendor bank details (STUB, admin only per SRS)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/vendors/{id}/commission": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors (Admin)"
                ],
                "summary": "Set per-vendor commission override (STUB, admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/vendors/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors (Admin)"
                ],
                "summary": "Reject vendor (STUB, admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/vendors/{id}/settle": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors (Admin)"
                ],
                "summary": "Trigger a payout settlement (STUB, admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/vendors/{id}/settlements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors (Admin)"
                ],
                "summary": "Get vendor settlement history (STUB, admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/vendors/{id}/suspend": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors (Admin)"
                ],
                "summary": "Suspend vendor (STUB, admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/vendors/{id}/wallet": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors (Admin)"
                ],
                "summary": "Get vendor wallet (STUB, admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/vendor/agreement/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Accept current vendor agreement version (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/vendor/documents/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Upload KYC document (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/vendor/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Get own vendor profile (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Update own vendor profile (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/vendor/staff": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "List vendor staff sub-accounts (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Create vendor staff sub-account (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/vendor/staff/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Update vendor staff sub-account (STUB)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Staff ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Delete vendor staff sub-account (STUB)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Staff ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_coolmate_ecommerce-backend_internal_utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Paste: `Bearer \u003caccessToken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}